Cleaned up print statements, unified command system
Previously, many commands required you to type in the console
This is now changed - the commands are in the function, but they don't
have a way to be called
They will probably be deleted in the cleanup, then some implemented for
the dungeon game.
As for print statements, there is a flag (that will be deleted in
cleanup) that decides whether the console is printed to or not, rather
than always printing to it.